+++ gradio/__init__.py
@@ -1,6 +1,7 @@
 import json
 
 import gradio.components as components
+import gradio.layouts as layouts
 import gradio.processing_utils
 import gradio.templates
 import gradio.themes as themes
