+++ gradio/cli/commands/__init__.py
@@ -1,5 +1,5 @@
 from .cli_env_info import print_environment_info
-from .create import main as custom_component
+from .components import app as custom_component
 from .deploy_space import deploy
 from .reload import main as reload
