+++ gradio/cli/commands/components/__init__.py
@@ -1,0 +1,3 @@
+from .app import app
+
+__all__ = ["app"]
