+++ gradio/layouts/column.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-
 from typing import Literal
 
 from gradio_client.documentation import document, set_documentation_group
@@ -9,8 +8,8 @@ from gradio.blocks import BlockContext
 from gradio.component_meta import ComponentMeta
 from gradio.deprecation import warn_deprecation
 
-
 set_documentation_group("layout")
+
 
 @document()
 class Column(BlockContext, metaclass=ComponentMeta):
