+++ gradio/layouts/row.py
@@ -8,7 +8,6 @@ from gradio.blocks import BlockContext
 from gradio.component_meta import ComponentMeta
 from gradio.deprecation import warn_style_method_deprecation
 
-
 set_documentation_group("layout")
